test: harden regression fixtures
