+++ crates/vize/tests/check_nuxt2_plugin_injections_cli.rs
@@ -102,227 +102,7 @@ fn check_nuxt2_false_positive_fixture_matrix() {
         return;
     };
     let project_root = create_project("nuxt2-false-positive-matrix");
-
-    write_file(
-        &project_root,
-        "vize.config.json",
-        r#"{
-  "vue": { "version": "2.7" },
-  "typeChecker": { "legacyVue2": true }
-}"#,
-    );
-    write_file(&project_root, "nuxt.config.ts", "export default {}\n");
-    write_file(
-        &project_root,
-        "tsconfig.json",
-        r##"{
-  "compilerOptions": {
-    "strict": true,
-    "target": "ES2022",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "baseUrl": ".",
-    "paths": { "~/*": ["*"] },
-    "noEmit": true
-  },
-  "include": ["components/**/*.vue", "pages/**/*.vue", "plugins/**/*.ts", "types/**/*.d.ts"]
-}"##,
-    );
-    write_file(
-        &project_root,
-        "types/nuxt.d.ts",
-        r##"declare module "@nuxt/types" {
-  export interface Context {}
-  export interface NuxtAppOptions {}
-}
-
-declare module "@nuxtjs/composition-api" {
-  export interface UseContextReturn
-    extends Omit<import("@nuxt/types").Context, "route" | "query" | "from" | "params"> {}
-  export function useContext(): UseContextReturn;
-}
-
-declare module "#app" {
-  export interface NuxtApp {}
-}
-"##,
-    );
-    write_file(
-        &project_root,
-        "plugins/logger.ts",
-        r#"export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
-  inject("logger", {
-    info(message: string) {
-      return message.length;
-    },
-  });
-};
-"#,
-    );
-    write_file(
-        &project_root,
-        "plugins/auth.ts",
-        r#"export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
-  inject("auth", {
-    loggedIn: true,
-    userName() {
-      return "Ada";
-    },
-  });
-};
-"#,
-    );
-    write_file(
-        &project_root,
-        "plugins/gtm.ts",
-        r#"export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
-  inject("gtm", {
-    push(event: { name: string }) {
-      return event.name;
-    },
-  });
-};
-"#,
-    );
-    write_file(
-        &project_root,
-        "plugins/repository.ts",
-        r#"export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
-  inject("accountRepository", {
-    find(id: string) {
-      return { id, label: id.toUpperCase() };
-    },
-  });
-};
-"#,
-    );
-    write_file(
-        &project_root,
-        "components/KeyboardPanel.vue",
-        r#"<script lang="ts">
-import { defineComponent, type PropType } from 'vue'
-
-export type Folder = { id: string; label: string }
-
-export default defineComponent({
-  props: {
-    folders: {
-      type: Array as PropType<Folder[]>,
-      required: true,
-    },
-  },
-  emits: {
-    'click-folder': (_folder: Folder) => true,
-    'input:math-key': (_key: string) => true,
-  },
-  methods: {
-    clickFirst() {
-      this.$emit('click-folder', this.folders[0])
-    },
-    inputKey() {
-      this.$emit('input:math-key', 'plus')
-    },
-  },
-})
-</script>
-
-<template>
-  <button @click="clickFirst">{{ folders[0].label }}</button>
-</template>
-"#,
-    );
-    write_file(
-        &project_root,
-        "components/OverlayPanel.vue",
-        r#"<script lang="ts">
-import { defineComponent } from 'vue'
-
-export default defineComponent({
-  emits: {
-    'update:is-opened-overlay-loading': (_value: boolean) => true,
-  },
-  methods: {
-    close() {
-      this.$emit('update:is-opened-overlay-loading', false)
-    },
-  },
-})
-</script>
-
-<template>
-  <button @click="close">Close</button>
-</template>
-"#,
-    );
-    write_file(
-        &project_root,
-        "pages/index.vue",
-        r#"<script lang="ts">
-import { defineComponent, type PropType } from 'vue'
-import { useContext } from '@nuxtjs/composition-api'
-import KeyboardPanel, { type Folder } from '~/components/KeyboardPanel.vue'
-import OverlayPanel from '~/components/OverlayPanel.vue'
-
-const componentProps = {
-  folders: {
-    type: Array as PropType<Folder[]>,
-    required: true,
-  },
-}
-
-function usePanelState() {
-  return {
-    panelTitle: 'Folders',
-    selectedId: 'root',
-  }
-}
-
-export default defineComponent({
-  components: { KeyboardPanel, OverlayPanel },
-  props: componentProps,
-  setup(props, { emit }) {
-    const context = useContext()
-    const repoItem = context.$accountRepository.find(props.folders[0].id)
-    context.$logger.info(context.$auth.userName())
-    context.$gtm.push({ name: repoItem.label })
-
-    const onClickFolder = (folder: Folder) => {
-      emit('click-folder', folder)
-      context.$logger.info(folder.label)
-    }
-    const onInputMathKey = (key: string) => {
-      emit('input:math-key', key)
-    }
-    const onUpdateIsOpenedOverlayLoading = (value: boolean) => {
-      emit('update:is-opened-overlay-loading', value)
-    }
-
-    return {
-      ...usePanelState(),
-      repoLabel: repoItem.label,
-      onClickFolder,
-      onInputMathKey,
-      onUpdateIsOpenedOverlayLoading,
-    }
-  },
-})
-</script>
-
-<template>
-  <section>
-    <h1>{{ panelTitle }} {{ selectedId }} {{ repoLabel }}</h1>
-    <KeyboardPanel
-      :folders="folders"
-      @click-folder="onClickFolder"
-      @input:math-key="onInputMathKey"
-    />
-    <OverlayPanel
-      @update:is-opened-overlay-loading="onUpdateIsOpenedOverlayLoading"
-    />
-  </section>
-</template>
-"#,
-    );
+    write_false_positive_fixture(&project_root);
 
     let output = Command::new(env!("CARGO_BIN_EXE_vize"))
         .current_dir(&project_root)
@@ -373,6 +153,61 @@ export default defineComponent({
     );
 
     let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[cfg(feature = "legacy")]
+const FALSE_POSITIVE_FIXTURES: &[(&str, &str)] = &[
+    (
+        "vize.config.json",
+        include_str!("fixtures/nuxt2_false_positive/vize.config.json"),
+    ),
+    (
+        "nuxt.config.ts",
+        include_str!("fixtures/nuxt2_false_positive/nuxt.config.ts"),
+    ),
+    (
+        "tsconfig.json",
+        include_str!("fixtures/nuxt2_false_positive/tsconfig.json"),
+    ),
+    (
+        "types/nuxt.d.ts",
+        include_str!("fixtures/nuxt2_false_positive/types/nuxt.d.ts"),
+    ),
+    (
+        "plugins/logger.ts",
+        include_str!("fixtures/nuxt2_false_positive/plugins/logger.ts"),
+    ),
+    (
+        "plugins/auth.ts",
+        include_str!("fixtures/nuxt2_false_positive/plugins/auth.ts"),
+    ),
+    (
+        "plugins/gtm.ts",
+        include_str!("fixtures/nuxt2_false_positive/plugins/gtm.ts"),
+    ),
+    (
+        "plugins/repository.ts",
+        include_str!("fixtures/nuxt2_false_positive/plugins/repository.ts"),
+    ),
+    (
+        "components/KeyboardPanel.vue",
+        include_str!("fixtures/nuxt2_false_positive/components/KeyboardPanel.vue"),
+    ),
+    (
+        "components/OverlayPanel.vue",
+        include_str!("fixtures/nuxt2_false_positive/components/OverlayPanel.vue"),
+    ),
+    (
+        "pages/index.vue",
+        include_str!("fixtures/nuxt2_false_positive/pages/index.vue"),
+    ),
+];
+
+#[cfg(feature = "legacy")]
+fn write_false_positive_fixture(project_root: &Path) {
+    for (path, contents) in FALSE_POSITIVE_FIXTURES {
+        write_file(project_root, path, contents);
+    }
 }
 
 fn create_project(name: &str) -> PathBuf {
