+++ crates/vize/tests/fixtures/nuxt2_false_positive/nuxt.config.ts
@@ -1,0 +1,1 @@
+export default {}
