+++ crates/vize/tests/fixtures/nuxt2_false_positive/plugins/auth.ts
@@ -1,0 +1,8 @@
+export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
+  inject("auth", {
+    loggedIn: true,
+    userName() {
+      return "Ada";
+    },
+  });
+};
