+++ crates/vize/tests/fixtures/nuxt2_false_positive/plugins/gtm.ts
@@ -1,0 +1,7 @@
+export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
+  inject("gtm", {
+    push(event: { name: string }) {
+      return event.name;
+    },
+  });
+};
