+++ crates/vize/tests/fixtures/nuxt2_false_positive/plugins/logger.ts
@@ -1,0 +1,7 @@
+export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
+  inject("logger", {
+    info(message: string) {
+      return message.length;
+    },
+  });
+};
