+++ crates/vize/tests/fixtures/nuxt2_false_positive/plugins/repository.ts
@@ -1,0 +1,7 @@
+export default (_context: unknown, inject: (key: string, value: unknown) => void) => {
+  inject("accountRepository", {
+    find(id: string) {
+      return { id, label: id.toUpperCase() };
+    },
+  });
+};
