+++ crates/vize/tests/fixtures/nuxt2_false_positive/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "target": "ES2022",
+    "module": "ESNext",
+    "moduleResolution": "bundler",
+    "baseUrl": ".",
+    "paths": { "~/*": ["*"] },
+    "noEmit": true
+  },
+  "include": ["components/**/*.vue", "pages/**/*.vue", "plugins/**/*.ts", "types/**/*.d.ts"]
+}
