+++ crates/vize_atelier_vapor/src/lib.rs
@@ -16,6 +16,8 @@ pub mod steps;
 mod tests;
 #[cfg(test)]
 mod tests_dotted_slots;
+#[cfg(test)]
+mod tests_setup_components;
 
 pub use compile::{
     VaporCompileResult, VaporCompilerOptions, compile_vapor, compile_vapor_with_diagnostics,
