+++ crates/vize_musea/src/parse.rs
@@ -576,41 +576,6 @@ defineArt("./base-button.vue", {
     }
 
     #[test]
-    fn test_parse_define_art_metadata_matrix() {
-        let allocator = Bump::new();
-        let source = r#"
-<script>
-export const localKind = "mixed";
-</script>
-
-<script setup>
-import { default as AliasButton } from "./AliasButton.vue";
-
-defineArt(AliasButton, {
-  title: "Alias Button",
-  category: "Components",
-  tags: ["alias", localKind],
-  status: "ready",
-});
-</script>
-
-<art>
-  <variant name="Primary" default>
-    <AliasButton>Primary</AliasButton>
-  </variant>
-</art>
-"#;
-
-        let desc = parse_art(&allocator, source, ArtParseOptions::default()).unwrap();
-
-        assert_eq!(desc.metadata.title, "Alias Button");
-        assert_eq!(desc.metadata.component, Some("./AliasButton.vue"));
-        assert_eq!(desc.metadata.category, Some("Components"));
-        assert_eq!(desc.metadata.tags.as_slice(), ["alias"]);
-        assert_eq!(desc.metadata.status, crate::types::ArtStatus::Ready);
-    }
-
-    #[test]
     fn test_extract_attr() {
         assert_eq!(extract_attr(r#"title="Hello""#, "title"), Some("Hello"));
         assert_eq!(extract_attr(r#"title='Hello'"#, "title"), Some("Hello"));
